Add separate package for prototype and registry design pattern codes
