+++ LLD/src/designpatterns/prototype/Prototype.java
@@ -1,4 +1,4 @@
-package designpatterns;
+package designpatterns.prototype;
 
 public interface Prototype<T> {
     T clone();
